Implement "start" and "privacy" commands.
Doing this took a significant refactor of the code. Until now, we always
assumed we would send one message per day and then immediately exit
until we were woken up again. This worked fine for a very long time, but
new Telegram policies[0] require the bot to provide users with a privacy
policy.

Because the bot no longer exits immediately, we don't need the systemd
service and timer files anymore.

There are still some TODOs left in the code, and the README still needs
to be updated.

[0]: https://t.me/BotNews/96