Make sending JSON parameter values more ergonomic.
This commit adds a new package, internal/mustjson, which provides an
Encode function that panics if json.Marshal returns an error. This makes
it easier to send JSON parameter values in the bot code, because we can
just use normal Go values directly and don't need to encode them by hand.

This approach has two downsides, one major and one minor. The major
downside is that mustjson.Encode is necessarily less efficient than
manual encoding, and it promotes the use of panics and globals.

The minor downside is that mustjson.Encode is not as general as it could
be. A function like:

    func Must[T any](v T, err error) T { ... }

(modeled after "html/template".Must) could be more generally applicable,
but it's less clear at the call site (because we have to convert the
result to a string when we use it with json.Marshal), and I can't think
of a good package name for it.