Fix "double-posting" bug.
When the bot received an update during the minute starting at 3:01 PM in
Amsterdam, it would post the energy report twice.

The bot now checks whether it's 3:01 PM in Amsterdam AND the previous
energy report was posted more than two minutes ago (this is arbitrary,
but it seems to work).