Replace uses of fp.Reduce with sliceutil.Min/Max.
slices.Min/Max panic when the slice is empty, which could happen if the
prices are not yet availble, which is unacceptable for us as it would
crash the bot. sliceutil.Min/Max returns positive or negative infinity,
respectively.

But really, we should just check if the slice is empty before calling
slices.Min/Max, but that's a change for another day.