Refactor datetime package to be more testable.
Tomorrow has been split into two pure functions: Tomorrow and Format,
which can be tested independently. The Amsterdam function has been
renamed to Now. It uses a package-level variable that is set to a
function that returns a time.Time, which can be overridden in tests.

As a drive-by change, I refactored GetGreeting to be a pure function
as well that takes a time.Time and returns a time-appropriate greeting.