Remove logic from template.
Now, the template doesn't contain any logic beyond a simple for-range
loop anymore, and the caller is responsible for providing the data the
template needs. This temporarily makes the callsite a bit more complex,
but will also make it easier to swap Templ (which has served me well)
for "html/template".