Improve performance of ranges.Format further.
According to benchstat[0], this change improves performance from 48.95
ns/op (±2%) to 45.38 ns/op (±1%), a 7.3% reduction in execution time
that is statistically significant.

The "more" benchmark, which I also ran on the old code, shows an
improvement from 156.6 ns/op (±1%) to 130.7 ns/op (±1%), representing
a statistically significant 16.54% reduction in execution time.

[0]: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat