Switch from Templ to "html/template". (#67)
This change unexpectedly increases the binary size by ~23% compared to
v1.0.59 (from ~8.5 MB to ~11 MB). It would be interesting to dig deeper
into this, but this cost is acceptable as it allows us to get rid of a
third-party dependency that, while it has served us well, needs to be
kept up-to-date.