Refactor EnergyPrices struct.
This commit ended up larger than I planned, but I think it made sense to
make the changes I made in one fell swoop.

I renamed the EnergyPrices struct to Prices, and moved it into its own
package, "internal/prices". I also removed the Get method, and replaced
it with the All method, which returns an iter.Seq2[int, float64] to
avoid exposing implementation details and prevent the possibility of
accessing hours that are out of bounds.

Closes #71.