Implement "-kickstart" flag.
When the "-kickstart" flag is passed to the binary, it immediately posts
an energy report and exits. The flag name is inspired by launctl
kickstart, which forces an existing service to start.

One downside is that this allows the combination "-kickstart -v", even
though printing the version and kickstarting the bot are mutually
exclusive operations. I've read that Cobra can handle this, but I'd
rather not pull it in. The current behavior (behaving as if only the
"-v" flag was passed) is fine.

Closes #74.