Change price overview to an expanded blockquote.
The collapsed blockquote was never a popular change, so I'm changing it
back after receiving feedback. The reason we still wrap the price
overview in a blockquote is that, at least on iOS, it looks better than
a code block outside of a blockquote. (Though this looks worse on macOS.
Ugh.)