Refactor job monitoring.
I implemented a new Monitor method on the "cronitor".Monitor struct,
which takes a function or closure and communicates its state to
Cronitor. This greatly simplifies the flow of the code in main, making
it much more readable and less bug-prone. As a nice bonus, it also made
it pretty easy to fix #81 (which would cause the bot to crash if the
template data was not yet available, and persisted for some time because
I didn't want to add yet more spaghetti complexity).