Refactor Telegram API client.
This commit adds a type-safe and very focused Telegram API client that
exposes just the methods and parameter combinations we need. This
improves both safety and readability. I also took this opportunity to
clean up when, where and what we log, which should hopefully make it
easier to figure out the cause of any problems when they arise.