Update API query parameters.
For reasons that are not entirely clear to me, the fromDate and tillDate
timestamps have been shifted by one hour. This bug was found by a very
attentive and friendly channel member.

Initially, I went down a deep rabbit hole thinking there must be a
problem with the "internal/prices" package, before finding this much
simpler but much more insidious root cause. The "prices" package could
be more scrupulous about rounding and such, but that's a change for
another day.