Pass stringified chat ID to privacy template.
When I refactored the code to use the ChatID type everywhere, I
accidentally broke the display of the user's ID in the privacy policy.
Even though it implements "fmt".Stringer, "html/template" doesn't seem
to call String() automatically, and so it displayed the addresses of the
two pointers in the struct (which isn't terribly illuminating to us
mortals).