Move state values to monitor.go.
Unfortunately, we can't make the state values constants, because the
the state transition test wants to take a pointer to them, which
constants do not allow. There are (ugly, but sometimes necessary) ways
to work around that[0], but that's overkill here. Since the states
aren't exported, keeping them as variables seems fine.

[0]: https://xeiaso.net/notes/2024/go-pointer-constant/