Add retry mechanism.
Recently, the energy prices haven't been available at 3:01 PM on several
occasions (perhaps because the auction takes longer?), causing
unpredictable failures. This has required us to manually retry the job,
which is inconvenient.

To fix this, this commit makes it so the CronJob runs every hour from
3:01 PM to 8:01 PM and writes a date-stamped file when the report has
been posted successfully. Subsequent runs check for the existence of
this file and exit early if it is present.