Restrict replies to users followed by the bot.
This is a fix for a bug where the bot stopped including the seemingly
required but empty "allow" array in a threadgate (to make it so no one
can reply), even though our code didn't change.

It's still a bit mysterious why this started happening or why Bluesky
accepted it even though it does the exact opposite of what we want, but
this fixes it since the bot only follows me.